backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
reportlab==4.4.1
sqlparse==0.5.3
typing_extensions==4.13.2
